+++ README.md
@@ -14,8 +14,9 @@ SkyWalking Extension Plugins for Scala 3
 | plugin              | library  | tested version | breakthrough points                                             |
 |---------------------|----------|----------------|-----------------------------------------------------------------|
 | caliban-v2x-plugin  | caliban  | 2.0.1          | `GraphQLInterpreter#executeRequest`                             |
-| zio-v2x-plugin      | zio      | 2.0.0          | `FiberRuntime#run`,`zio.Executor`,`zio.internal.ZScheduler`     |
-| ziogrpc-v06x-plugin | zio-grpc | 0.6.0-test4    | `ZChannel#newCall`,`ZServerCallHandler#startCall`,`ZServerCall` |
+| zio-v200-plugin     | zio      | 2.0.0 ~ 2.0.2  | `FiberRuntime#run`,`zio.Executor`,`zio.internal.ZScheduler`     |
+| zio-v203-plugin     | zio      | 2.0.3 ~ -      | `FiberRuntime#run`,`zio.Executor`,`zio.internal.ZScheduler`     |
+| ziogrpc-v06x-plugin | zio-grpc | 0.6.0-RC5      | `ZChannel#newCall`,`ZServerCallHandler#startCall`,`ZServerCall` |
 | ziohttp-v2x-plugin  | zio-http | 2.0.0-RC10     | `Http.collectHttp`                                              |
 
 > Other small versions of the library supported by this plugin may also work, but they have not been tested.
@@ -23,15 +24,15 @@ SkyWalking Extension Plugins for Scala 3
 Also need `apm-jdk-threadpool-plugin-x.y.z.jar` and `apm-jdk-threading-plugin-x.y.z.jar`.
 
 ## Available Configurations
-| key                                           | description                                                                           |
-|-----------------------------------------------|---------------------------------------------------------------------------------------|
-| `plugin.caliban.url_prefix`                   | Add a custom prefix to the graphql operation, default is `GraphQL/`.                  |
-| `plugin.caliban.ignore_url_prefixes`          | Ignore operation name that start with this prefix, i.e. no span will be created.      |
-| `plugin.caliban.collect_variables`            | Collect request variables.                                                            |
-| `plugin.caliban.variables_length_threshold`   | How many characters to keep and send to the OAP backend.                              |
-| `plugin.ziohttp.ignore_url_prefixes`          | Ignore request path that should start with this prefix, i.e. no span will be created. |
-| `plugin.ziohttp.collect_http_params`          | Collect http query params.                                                            |
-| `plugin.ziohttp.http_params_length_threshold` | How many characters to keep and send to the OAP backend.                              |
+| key                                             | description                                                                           |
+|-------------------------------------------------|---------------------------------------------------------------------------------------|
+| `plugin.caliban.url_prefix`                     | Add a custom prefix to the graphql operation, default is `Caliban/GraphQL/`.          |
+| `plugin.caliban.ignore_url_prefixes`            | Ignore operation name that start with this prefix, i.e. no span will be created.      |
+| `plugin.caliban.collect_variables`              | Collect request variables.                                                            |
+| `plugin.caliban.variables_length_threshold`     | How many characters to keep and send to the OAP backend.                              |
+| `plugin.ziohttpv2.ignore_url_prefixes`          | Ignore request path that should start with this prefix, i.e. no span will be created. |
+| `plugin.ziohttpv2.collect_http_params`          | Collect http query params.                                                            |
+| `plugin.ziohttpv2.http_params_length_threshold` | How many characters to keep and send to the OAP backend.                              |
 
 The prefix should be added when passing command line parameters, such as: `-Dskywalking.plugin.caliban.url_prefix=GQL/`
