+++ zio-v200-plugin/src/main/resources/skywalking-plugin.def
@@ -1,0 +1,2 @@
+zio-2.0.0-fiber=org.bitlap.skywalking.apm.plugin.zio.v200.define.ZioFiberRuntimeInstrumentation
+zio-2.0.0-executor=org.bitlap.skywalking.apm.plugin.zio.v200.define.ZioExecutorInstrumentation
