+++ zio-v203-plugin/src/main/resources/skywalking-plugin.def
@@ -1,0 +1,2 @@
+zio-2.0.3-fiber=org.bitlap.skywalking.apm.plugin.zio.v203.define.ZioFiberRuntimeInstrumentation
+zio-2.0.3-executor=org.bitlap.skywalking.apm.plugin.zio.v203.define.ZioExecutorInstrumentation
