--- zio-v2x-plugin/src/main/resources/skywalking-plugin.def
@@ -1,2 +1,0 @@
-zio-2.x-fiber=org.bitlap.skywalking.apm.plugin.zio.define.ZioFiberRuntimeInstrumentation
-zio-2.x-executor=org.bitlap.skywalking.apm.plugin.zio.define.ZioExecutorInstrumentation
