--- zio-v2x-plugin/src/main/scala/org/bitlap/skywalking/apm/plugin/zio/ExecutorType.scala
@@ -1,12 +1,0 @@
-package org.bitlap.skywalking.apm.plugin.zio
-
-/** @author
- *    梦境迷离
- *  @version 1.0,2023/6/9
- */
-enum ExecutorType(val classNamePrefix: String):
-  case Executor                extends ExecutorType("zio.Executor$$anon$")
-  case ZScheduler              extends ExecutorType("zio.internal.ZScheduler")
-  case DefaultBlockingExecutor extends ExecutorType("zio.internal.DefaultExecutors$")
-
-end ExecutorType
